Package Internals

Documentation of the OpenEphysLoader.jl internals.

OpenEphysLoader.BlockBuffer — Type.

Type to buffer continuous file contents

source

OpenEphysLoader.BlockHeader — Type.

Represents the header of each data block

source

OpenEphysLoader.DataBlock — Type.

Represents the entirety of a data block

source

OpenEphysLoader.MATLABdata — Type.

Abstract class for representing matlab code fragments

source

OpenEphysLoader.MATfloat — Type.

type for representing Matlab floatingpoint numbers

source

OpenEphysLoader.MATint — Type.

Type for representing Matlab integers

source

OpenEphysLoader.MATstr — Type.

Type for representing Matlab strings

source

OpenEphysLoader.OESignalTree — Type.

OESignalTree{T<:OEProcessor}(chain_e::LightXML.XMLElement, [recording_anmes::Set])

Signal tree for recording processors. Since OpenEphysLoader currently on works on .continuous file types, this will search for the first OERhythmProcessor and make a signal tree up to that point.

Construct with a XML signalchain element, and a set of processor names that are valid recording nodes.

See Tree for field information.

source

OpenEphysLoader.SignalNode — Type.

SignalNode{T<:OEProcessor}

Node type for OEProcessor signal chain, subtype of TreeNode.

See TreeNode for information on fields.

source

OpenEphysLoader.Tree — Type.

Tree{T}

Abstract type for tree structure, with type T content.

Contains a group of TreeNode in the single required field:

Required Fields

nodes Indexable list of TreeNode elements.

source

OpenEphysLoader.TreeNode — Type.

TreeNode{T}

Abstract node type for tree structure, with type T content.

Subtypes must have the following fields:

Required Fields

content T content of node.

parent Int ID of parent node

children Vector{Int} IDs of children node

source

OpenEphysLoader.add_continuous_meta! — Method.

Add data from Continuous_Data.openephys to OESettings from settings.xml

source

OpenEphysLoader.channel_arr — Function.

Parse XML Element PROCESSOR and recover channel metadata.

source

OpenEphysLoader.check_filesize — Method.

Check that file could be comprised of header and complete data blocks

source

OpenEphysLoader.convert_block! — Method.

Convert the wacky data format in OpenEphys continuous files

source

OpenEphysLoader.find_matching_proc — Method.

Find id of processor in OESignalTree that matches id of XML processor element

source

OpenEphysLoader.matread — Method.

read a Matlab source line

source

OpenEphysLoader.parseline — Function.

Parse a line of Matlab source code

source

OpenEphysLoader.parseto — Function.

Convert a string to the desired type

source

OpenEphysLoader.prepare_block! — Method.

Load data block if necessary

source

OpenEphysLoader.read_into! — Function.

Read file data block into data block buffer

source

OpenEphysLoader.read_into! — Method.

Read block header into header buffer

source

OpenEphysLoader.seek_to_block — Method.

Move io to data block

source

OpenEphysLoader.verify_tail! — Method.

Verify end of block marker

source