Public Documentation

Documentation for exported functions and types for OpenEphysLoader.jl

OpenEphysLoader — Module.

Module to read the binary data files created by the OpenEphys GUI

Provides array interfaces to file contents, without loading the entire file into memory

source

Array types

All array types are subtypes of the abstract type OEArray, and data from continuous files are subtypes of the abstract type OEContArray.

OpenEphysLoader.OEArray — Type.

Abstract array for file-backed OpenEphys data.

All subtypes support a ready-only array interface and should be constructable with a single IOStream argument.

source

OpenEphysLoader.OEContArray — Type.

Abstract array for file-backed continuous OpenEphys data.

Will throw CorruptedException if the data file has a corrupt OriginalHeader, is not the correct size for an .continuous file, or contains corrupt data blocks.

Subtype of abstract type OEArray are read only, and have with the following fields:

Fields

contfile ContinuousFile for the current file.

block buffer object for the data blocks in the file.

blockno the current block being access in the file.

check Bool to check each data block's validity.

source

The following array types can be used to access different aspects of the data:

OpenEphysLoader.SampleArray — Type.

SampleArray(type::Type{T}, io::IOStream, [check::Bool])

Subtype of OEContArray to provide file backed access to OpenEphys sample values. If type is a floating point type, then the sample value will be converted to voltage (in uV). Otherwise, the sample values will remain the raw ADC integer readings.

source

OpenEphysLoader.TimeArray — Type.

TimeArray(type::Type{T}, io::IOStream, [check::Bool])

Subtype of OEContArray to provide file backed access to OpenEphys time stamps. If type is a floating point type, then the time stamps will be converted to seconds. Otherwise, the time stamp will be the sample number.

source

OpenEphysLoader.RecNoArray — Type.

RecNoArray(type::Type{T}, io::IOStream, [check::Bool])

Subtype of OEContArray to provide file backed access to OpenEphys numbers.

source

Alternatively, all three aspects can be accessed simultaneously:

OpenEphysLoader.JointArray — Type.

JointArray(type::Type{T}, io::IOStream, [check::Bool])

Subtype of OEContArray to provide file backed access to OpenEphys data. Returns a tuple of type type, whose values represent (samplevalue, timestamp, recordingnumber). For a description of each, see SampleArray, TimeArray, and RecNoArray, respectively.

source

Information types

The following types provide information about OpenEphys files

OpenEphysLoader.OriginalHeader — Type.

OriginalHeader(io::IOStream)

Data in the header of binary OpenEphys files.

Will throw CorruptedException if header is corrupt, not an "OpenEphys" data format, or not version 0.4 of the data format.

Fields

format is the name of the data format.

version is the version number of the data format.

headerbytes is the number of bytes in the header.

description is a description of the header.

created is the date and time the file was created.

channel is the name of the channel used to acquire this data.

channeltype is the type of channel used to acquire this data.

samplerate is the sampling rate in Hz.

blocklength is the length in bytes of each block of data within the file.

buffersize is the size of the buffer used during acquisition, in bytes.

bitvolts are the Volts per ADC bit.

source

OpenEphysLoader.ContinuousFile — Type.

ContinuousFile(io::IOStream)

Type for an open continuous file.

Fields

io IOStream object.

nsample number of samples in a file.

nblock number of data blocks in a file.

header OriginalHeader of the current file.

source

Exceptions

OpenEphysLoader.CorruptedException — Type.

Exception type to indicate a malformed data file

source